finish v1 frontend

+++ frontend/js/login.js
@@ -1,4 +1,4 @@
-function submitForm() {
+function login() {
     let URL = "http://127.0.0.1:8000";
     let name = document.getElementById("name").value;
     let xhr = new XMLHttpRequest();
@@ -8,14 +8,38 @@ function submitForm() {
     xhr.send(requestBody);
     xhr.onreadystatechange = function() {
         if (xhr.readyState == XMLHttpRequest.DONE) {
-        if (xhr.status == 200) {
-            user = JSON.parse(xhr.responseText);
-            localStorage.setItem("name_user", user.name);
-            window.location = "./messages.html";
-        } else {
-            alert("Usuário não encontrado!");
-            localStorage.clear()
+            if (xhr.status == 200) {
+                user = JSON.parse(xhr.responseText);
+                localStorage.setItem("name_user", user.name);
+                window.location = "./messages.html";
+            } else {
+                alert("Usuário não encontrado!");
+                localStorage.clear()
+            }
         }
+    };
+}
+
+
+function register() {
+    let URL = "http://127.0.0.1:8000/register";
+    let name = document.getElementById("name").value;
+    let xhr = new XMLHttpRequest();
+    xhr.open("POST", URL, true);
+    xhr.setRequestHeader("Content-Type", "application/json");
+    let requestBody = JSON.stringify({"name": name});
+    xhr.send(requestBody);
+    xhr.onreadystatechange = function() {
+        if (xhr.readyState == XMLHttpRequest.DONE) {
+            if (xhr.status == 200) {
+                user = JSON.parse(xhr.responseText);
+                localStorage.setItem("name_user", user.name);
+                alert("Usuário cadastrado com sucesso!");
+                window.location = "./messages.html";
+            } else {
+                alert("Erro ao cadastrar usuário!");
+                localStorage.clear()
+            }
         }
     };
 }
